background pic started
I chose a random dark nature pic from google.  The idea is to have a themed selection to choose from that doesn't conflict with the buttons visibility.  Or there could be a new picture every day, being rotated, which we can get free pictures from unspash or whatever it's called.  This current picture is not sized properly and isn't fitting correctly on screen yet, so this needs to be fixed.