dressed up profile page
I cleaned up the profile page a bit.  I'd still like to make some minor changes, particularly with the add button.  I also got rid of the withdraw button and the sign out button and placed them into a separate file to grab from later.  My thinking is that we could put those buttons in the area where you click on your name and where you can change your profile picture.  We want to keep the money in the app for as long as possible until we start our own bank, so I think making an extra step to withdraw funds would help keep the money in the app for longer.